+++ nexus/users/models.py
@@ -33,42 +33,47 @@ class UserManager(BaseUserManager):
 
 class User(AbstractBaseUser, UUIDModel, PermissionsMixin):
 
-    GENDER = Choices(
-        ('MALE', 'M', _('male')),
-        ('FEMALE', 'F', _('female')),
-        ('OTHERS', 'O', _('others')),
+    GENDER_CHOICES = Choices(
+        ('NOT_SELECTED', 'NA', _('Not Selected')),
+        ('MALE', 'M', _('Male')),
+        ('FEMALE', 'F', _('Female')),
+        ('OTHERS', 'O', _('Others')),
     )
 
-    TSHIRT_SIZE = Choices(
-        ('SMALL', 'S', _('small')),
-        ('MEDIUM', 'M', _('medium')),
-        ('LARGE', 'L', _('large')),
-        ('EXTRA_LARGE', 'XL', _('extra_large')),
-        ('DOUBLE_EXTRA_LARGE', 'XXL', _('double extra large')),
+    TSHIRT_SIZE_CHOICES = Choices(
+        ('NOT_SELECTED', 'NA', _('Not Selected')),
+        ('SMALL', 'S', _('Small')),
+        ('MEDIUM', 'M', _('Medium')),
+        ('LARGE', 'L', _('Large')),
+        ('EXTRA_LARGE', 'XL', _('Extra Large')),
+        ('DOUBLE_EXTRA_LARGE', 'XXL', _('Double Extra Large')),
     )
 
-    first_name = models.CharField(_('First Name'), max_length=120, blank=True)
-    last_name = models.CharField(_('Last Name'), max_length=120, blank=True)
+    first_name = models.CharField(_('First Name'), null=False, blank=True, max_length=120)
+    last_name = models.CharField(_('Last Name'), null=False, blank=True, max_length=120)
     # https://docs.djangoproject.com/en/1.11/ref/contrib/postgres/fields/#citext-fields
-    email = CIEmailField(_('email address'), unique=True, db_index=True)
-    is_staff = models.BooleanField(_('staff status'), default=False,
+    email = CIEmailField(_('Email Address'), unique=True, db_index=True)
+    is_staff = models.BooleanField(_('Staff Status'), default=False,
                                    help_text='Designates whether the user can log into this admin site.')
 
-    is_active = models.BooleanField('active', default=True,
+    is_active = models.BooleanField('Active', default=True,
                                     help_text='Designates whether this user should be treated as '
                                               'active. Unselect this instead of deleting accounts.')
-    date_joined = models.DateTimeField(_('date joined'), default=timezone.now)
+    date_joined = models.DateTimeField(_('Date Of Joining'), default=timezone.now, null=False, blank=False)
+
+    gender = models.CharField(_('Gender'), default=GENDER_CHOICES.NOT_SELECTED, choices=GENDER_CHOICES, null=False,
+                              blank=False, max_length=6)
+    tshirt_size = models.CharField(_('Tshirt Size'), default=TSHIRT_SIZE_CHOICES.NOT_SELECTED,
+                                   choices=TSHIRT_SIZE_CHOICES, null=False, blank=False, max_length=10)
+    phone_number = PhoneNumberField(_('Phone Number'), default='', null=False, blank=True, max_length=13)
+    ticket_id = models.CharField(_('Ticket Id'), default=_('Not assigned'), null=False, blank=False, max_length=32)
+    is_core_organizer = models.BooleanField(_('Core Organizer Status'), default=False, null=False, blank=True,
+                                            help_text='Designates whether this user is a Core Organizer')
+    is_volunteer = models.BooleanField(_('Volunteer Status'), default=False, null=False, blank=True,
+                                       help_text='Designates whether this user is a Volunteer')
 
     USERNAME_FIELD = 'email'
     objects = UserManager()
-
-    gender = models.CharField(_('Gender'), default='O', choices=GENDER, max_length=6)
-    tshirt_size = models.CharField(_('Tshirt Size'), default='L', choices=TSHIRT_SIZE, max_length=10)
-    phone_number = PhoneNumberField(_('Phone Number'), blank=False, null=True, max_length=13)
-    ticket_id = models.CharField(_('Ticket Id'), default=_('Not assigned'), blank=False, null=False, max_length=32)
-    is_core_organizer = models.BooleanField(_('core organizer status'), default=False,
-                                            help_text='Is the user a core organizer')
-    is_volunteer = models.BooleanField(_('volunteer status'), default=False, help_text='Is the user a volunteer')
 
     class Meta:
         verbose_name = _('user')