+++ nexus/users/serializers.py
@@ -9,3 +9,5 @@ class UserSerializer(serializers.ModelSerializer):
         model = models.User
         fields = ['id', 'first_name', 'last_name', 'email', 'gender', 'tshirt_size', 'ticket_id', 'phone_number',
                   'is_core_organizer', 'is_volunteer', 'date_joined', 'is_active', 'is_staff', 'is_superuser']
+        read_only_fields = ['id', 'ticket_id', 'is_core_organizer', 'is_volunteer', 'date_joined', 'is_active',
+                            'is_staff', 'is_superuser']