+++ nexus/users/tests/test_models.py
@@ -3,13 +3,14 @@ from django.test import TestCase
 
 # nexus Stuff
 from nexus.users.models import User
+from tests import factories as f
 
 
 class UserModelTestCase(TestCase):
 
     def test_create_user(self):
-        u = User.objects.create_user(
-            email='f@F.com', password='abc', first_name="F", last_name='B',
+        u = f.create_user(
+            email='f@f.com', password='abc', first_name="F", last_name='B',
             gender='O', tshirt_size='L', phone_number='+910123456789'
         )
         assert u.is_active is True
@@ -19,7 +20,7 @@ class UserModelTestCase(TestCase):
         assert u.get_full_name() == 'F B'
         assert u.get_short_name() == 'F'
         assert str(u) == str(u.id)
-        assert u.gender == 'O'
+        assert u.gender == User.GENDER_CHOICES.OTHERS
         assert u.tshirt_size == 'L'
         assert u.is_core_organizer is False
         assert u.is_volunteer is False